Avoid float to long conversions.

+++ ojgl/src/utility/Duration.h
@@ -20,9 +20,6 @@ public:
     template <typename T = float>
     T toMilliseconds() const { return static_cast<T>(this->_duration); }
 
-    //  template <typename T = float>
-    //  T toNanoseconds() const { return std::chrono::duration_cast<std::chrono::duration<T, std::nano>>(_duration).count(); }
-
     Duration operator-() const { return Duration(-this->_duration); }
     Duration operator-(const Duration& other) const { return Duration(this->_duration - other._duration); }
     Duration operator+(const Duration& other) const { return Duration(this->_duration + other._duration); }
@@ -35,12 +32,9 @@ public:
     Duration& operator-=(const Duration& other);
     Duration& operator+=(const Duration& other);
 
-    //friend std::ostream& operator<<(std::ostream& left, const Duration& right);
-
 public:
-    static Duration seconds(double s) { return Duration(static_cast<long>(s * 1000)); }
+    static Duration seconds(long s) { return Duration(s * 1000); }
     static Duration milliseconds(long ms) { return Duration(ms); }
-    //static Duration nanoseconds(long long ns) { return Duration(std::chrono::nanoseconds(ns)); }
     static Duration maximum();
 
 private:
